finish Query, NameMethod, NamedNativeQuery for SecondRound

+++ SourceCode/SpringBoot/springBoot_secondRound/src/main/java/com/axonactive/homeSpringBoot/service/dto/CertificateOfPilotCanOnlyFly3AircraftsDTO.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class CertificateOfPilotCanOnlyFly3AircraftsDTO {
     private String employeeId;
 
-    private int totalAircraftsCanFly;
+    private Long totalAircraftsCanFly;
 }
